added anotehr test to common
