+++ tests/estimators/classification/test_common_deeplearning.py
@@ -4,7 +4,10 @@ import numpy as np
 from os import listdir, path
 import pytest
 import tempfile
+from tensorflow.keras.callbacks import LearningRateScheduler
 import warnings
+
+from art.defences.preprocessor import FeatureSqueezing, JpegCompression, SpatialSmoothing
 
 logger = logging.getLogger(__name__)
 
@@ -105,6 +108,51 @@ def test_shapes(get_default_mnist_subset, get_image_classifier_list):
     except NotImplementedError as e:
         warnings.warn(UserWarning(e))
 
+
+@pytest.mark.skipMlFramework("mxnet", "tensorflow","scikitlearn", "pytorch")
+def test_fit_kwargs(get_image_classifier_list, get_default_mnist_subset, default_batch_size):
+
+    (x_train_mnist, y_train_mnist), (_, _) = get_default_mnist_subset
+
+    # Test a valid callback
+    classifier, _ = get_image_classifier_list(one_classifier=True, from_logits=True)
+    kwargs = {"callbacks": [LearningRateScheduler(0.01)]}
+    classifier.fit(x_train_mnist, y_train_mnist, batch_size=default_batch_size, nb_epochs=1, **kwargs)
+
+    # Test failure for invalid parameters
+    kwargs = {"epochs": 1}
+    with pytest.raises(TypeError) as exception:
+        classifier.fit(x_train_mnist, y_train_mnist, batch_size=default_batch_size, nb_epochs=1, **kwargs)
+
+    assert "multiple values for keyword argument" in str(exception)
+
+
+def test_defences_predict(get_default_mnist_subset, get_image_classifier_list_defended, get_image_classifier_list):
+    (_, _), (x_test_mnist, y_test_mnist) = get_default_mnist_subset
+
+    classifier, _ = get_image_classifier_list_defended(one_classifier=True,
+                                                       defenses=["FeatureSqueezing", "JpegCompression",
+                                                                 "SpatialSmoothing"])
+    if classifier is not None:
+        assert len(classifier.preprocessing_defences) == 3
+
+        predictions_classifier = classifier.predict(x_test_mnist)
+
+        # Apply the same defences by hand
+        x_test_defense = x_test_mnist
+        clip_values = (0, 1)
+        fs = FeatureSqueezing(clip_values=clip_values, bit_depth=2)
+        x_test_defense, _ = fs(x_test_defense, y_test_mnist)
+        jpeg = JpegCompression(clip_values=clip_values, apply_predict=True)
+        x_test_defense, _ = jpeg(x_test_defense, y_test_mnist)
+        smooth = SpatialSmoothing()
+        x_test_defense, _ = smooth(x_test_defense, y_test_mnist)
+        # classifier, _ = get_image_classifier_list(one_classifier=True, from_logits=True)
+        classifier, _ = get_image_classifier_list(one_classifier=True)
+        predictions_check = classifier._model.predict(x_test_defense)
+
+        # Check that the prediction results match
+        np.testing.assert_array_almost_equal(predictions_classifier, predictions_check, decimal=4)
 
 # Note: because mxnet only supports 1 concurrent version of a model if we fit that model, all expected values will
 # change for all other tests using that fitted model
