+++ tests/estimators/classification/test_keras_tf.py
@@ -84,37 +84,6 @@ def test_learning_phase(get_image_classifier_list):
 
 
 @pytest.mark.only_with_platform("kerastf")
-def test_defences_predict(get_default_mnist_subset, get_image_classifier_list_defended, get_image_classifier_list):
-    clip_values = (0, 1)
-
-    fs = FeatureSqueezing(clip_values=clip_values, bit_depth=2)
-    jpeg = JpegCompression(clip_values=clip_values, apply_predict=True)
-    smooth = SpatialSmoothing()
-
-    (_, _), (x_test_mnist, y_test_mnist) = get_default_mnist_subset
-
-    classifier, _ = get_image_classifier_list_defended(one_classifier=True, from_logits=True,
-                                                       defenses=["FeatureSqueezing", "JpegCompression",
-                                                                 "SpatialSmoothing"])
-
-    assert len(classifier.preprocessing_defences) == 3
-
-    predictions_classifier = classifier.predict(x_test_mnist)
-
-    # Apply the same defences by hand
-    x_test_defense = x_test_mnist
-    x_test_defense, _ = fs(x_test_defense, y_test_mnist)
-    x_test_defense, _ = jpeg(x_test_defense, y_test_mnist)
-    x_test_defense, _ = smooth(x_test_defense, y_test_mnist)
-
-    classifier, _ = get_image_classifier_list(one_classifier=True, from_logits=True)
-    predictions_check = classifier._model.predict(x_test_defense)
-
-    # Check that the prediction results match
-    np.testing.assert_array_almost_equal(predictions_classifier, predictions_check, decimal=4)
-
-
-@pytest.mark.only_with_platform("kerastf")
 def test_pickle(get_image_classifier_list_defended, tmp_path):
     from art.defences.preprocessor import FeatureSqueezing
 
@@ -132,26 +101,6 @@ def test_pickle(get_image_classifier_list_defended, tmp_path):
     assert classifier._use_logits == loaded._use_logits
     assert classifier._input_layer == loaded._input_layer
     assert isinstance(loaded.preprocessing_defences[0], FeatureSqueezing)
-
-
-@pytest.mark.only_with_platform("kerastf")
-def test_fit_kwargs(get_image_classifier_list, get_default_mnist_subset, default_batch_size):
-    def get_lr(_):
-        return 0.01
-
-    (x_train_mnist, y_train_mnist), (_, _) = get_default_mnist_subset
-
-    # Test a valid callback
-    classifier, _ = get_image_classifier_list(one_classifier=True, from_logits=True)
-    kwargs = {"callbacks": [LearningRateScheduler(get_lr)]}
-    classifier.fit(x_train_mnist, y_train_mnist, batch_size=default_batch_size, nb_epochs=1, **kwargs)
-
-    # Test failure for invalid parameters
-    kwargs = {"epochs": 1}
-    with pytest.raises(TypeError) as exception:
-        classifier.fit(x_train_mnist, y_train_mnist, batch_size=default_batch_size, nb_epochs=1, **kwargs)
-
-    assert "multiple values for keyword argument" in str(exception)
 
 
 @pytest.mark.only_with_platform("kerastf")
